Temporarily Revert fix for #2546
This reverts commit 10125a946fd089a3fc259ca4820854fdb33ac2e0.
