+++ test/Microsoft.Dnx.CommonTestUtils/project.json
@@ -1,23 +1,22 @@
 {
-    "dependencies": {
-        "Microsoft.Dnx.Runtime": "1.0.0-*",
-        "Microsoft.Dnx.Runtime.Sources": "1.0.0-*",
-        "xunit.runner.aspnet": "2.0.0-aspnet-*",
-        "Newtonsoft.Json": "6.0.6"
-    },
+  "dependencies": {
+    "Microsoft.Dnx.Runtime": "1.0.0-*",
+    "Microsoft.Dnx.Runtime.Sources": "1.0.0-*",
+    "xunit.runner.aspnet": "2.0.0-aspnet-*"
+  },
 
-    "frameworks": {
-        "dnx451": {
-            "frameworkAssemblies": {
-                "System.IO.Compression.FileSystem": ""
-            }
-        },
-        "dnxcore50": {
-            "dependencies": {
-                "System.IO.Compression.ZipFile": "4.0.0-*"
-            }
-        }
+  "frameworks": {
+    "dnx451": {
+      "frameworkAssemblies": {
+        "System.IO.Compression.FileSystem": ""
+      }
     },
+    "dnxcore50": {
+      "dependencies": {
+        "System.IO.Compression.ZipFile": "4.0.0-*"
+      }
+    }
+  },
 
     "commands": {
         "test": "xunit.runner.aspnet"
